Added DocStrings
