+++ pom.xml
@@ -36,13 +36,6 @@
 			<artifactId>spring-boot-starter-security</artifactId>
 			<version>3.1.5</version>
 		</dependency>
-<!--		&lt;!&ndash; https://mvnrepository.com/artifact/io.jsonwebtoken/jjwt-api &ndash;&gt;-->
-<!--		<dependency>-->
-<!--			<groupId>io.jsonwebtoken</groupId>-->
-<!--			<artifactId>jjwt-api</artifactId>-->
-<!--			<version>0.11.2</version>-->
-<!--		</dependency>-->
-
 		<!-- https://mvnrepository.com/artifact/org.projectlombok/lombok -->
 		<dependency>
 			<groupId>org.projectlombok</groupId>
