+++ src/main/java/com/kirana/Kirana_Register/KiranaRegisterApplication.java
@@ -13,23 +13,16 @@ import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
+/**
+ * Main class for the Kirana Register application.
+ */
 @SpringBootApplication
-//@EnableSwagger2
 public class KiranaRegisterApplication {
-
-//	@Bean
-//	public Docket api() {
-//		return new Docket(DocumentationType.SWAGGER_2).select()
-//				.apis(RequestHandlerSelectors.withClassAnnotation(RestController.class)).paths(PathSelectors.any())
-//				.build().apiInfo(apiInfo()).useDefaultResponseMessages(false);
-//	}
-//	@Bean
-//	public ApiInfo apiInfo() {
-//		final ApiInfoBuilder builder = new ApiInfoBuilder();
-//		return builder.build();
-//	}
-
-
+	/**
+	 * Main method to start the Spring Boot application.
+	 *
+	 * @param args Command-line arguments.
+	 */
 	public static void main(String[] args) {
 		SpringApplication.run(KiranaRegisterApplication.class, args);
 	}
