+++ src/main/java/com/kirana/Kirana_Register/config/RedisConfig.java
@@ -15,9 +15,19 @@ import javax.cache.CacheManager;
 import javax.cache.Caching;
 import javax.cache.spi.CachingProvider;
 
+/**
+ * Configuration class for Redis integration in Kirana Register application.
+ * Configures the Redisson client, creates a custom CacheManager using Redisson,
+ * and provides a JCacheProxyManager for Bucket4j rate limiting.
+ */
 @Configuration
 public class RedisConfig
 {
+    /**
+     * Creates and configures a Redisson client.
+     *
+     * @return Configured Redisson client instance.
+     */
     @Bean
     public Config config()
     {
@@ -26,6 +36,12 @@ public class RedisConfig
         return config;
     }
 
+    /**
+     * Creates a custom CacheManager using Redisson for caching.
+     *
+     * @param config Redisson configuration for cache management.
+     * @return Configured CacheManager instance.
+     */
     @Bean(name="my_cache_manager")
     public CacheManager cacheManager(Config config) {
         CacheManager manager = Caching.getCachingProvider().getCacheManager();
@@ -33,6 +49,12 @@ public class RedisConfig
         return manager;
     }
 
+    /**
+     * Creates a JCacheProxyManager for Bucket4j rate limiting using the custom CacheManager.
+     *
+     * @param cacheManager Configured CacheManager instance.
+     * @return JCacheProxyManager for Bucket4j rate limiting.
+     */
     @Bean
     ProxyManager<String> proxyManager(CacheManager cacheManager) {
         return new JCacheProxyManager<>(cacheManager.getCache("cache"));
@@ -42,6 +64,13 @@ public class RedisConfig
      * reference:
      * https://github.com/MarcGiffing/bucket4j-spring-boot-starter/issues/73
      * */
+    /**
+     * Provides a SyncCacheResolver for Bucket4j rate limiting using the custom CacheManager.
+     * This resolver is marked as Primary to be used as the primary cache resolver.
+     *
+     * @param cacheManager Configured CacheManager instance.
+     * @return JCacheCacheResolver for Bucket4j rate limiting.
+     */
     @Bean
     @Primary
     public SyncCacheResolver bucket4jCacheResolver(CacheManager cacheManager) {
