--- src/main/java/com/kirana/Kirana_Register/dump/ExchangeRateResponse.java
@@ -1,51 +1,0 @@
-package com.kirana.Kirana_Register.dump;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import java.util.Map;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ExchangeRateResponse {
-
-    private String base;
-    private String date;
-    private Map<String, Double> rates;
-
-    public ExchangeRateResponse() {
-    }
-
-    @Override
-    public String toString() {
-        return "ExchangeRateResponse{" +
-                "base='" + base + '\'' +
-                ", date='" + date + '\'' +
-                ", rates=" + rates +
-                '}';
-    }
-
-    public String getBase() {
-        return base;
-    }
-
-    public void setBase(String base) {
-        this.base = base;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public void setDate(String date) {
-        this.date = date;
-    }
-
-    public Map<String, Double> getRates() {
-        return rates;
-    }
-
-    public void setRates(Map<String, Double> rates) {
-        this.rates = rates;
-    }
-
-    // getters and setters
-}
