+++ src/main/java/com/kirana/Kirana_Register/entities/Transaction.java
@@ -1,26 +1,48 @@
 package com.kirana.Kirana_Register.entities;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.util.Date;
 
+/**
+ * Entity class representing a financial transaction.
+ * This class is annotated with @Document to specify its storage in MongoDB.
+ */
 @Document(collection= "transaction")
 public class Transaction {
-//    @JsonProperty("transaction_id")
+    /**
+     * Unique identifier for the transaction.
+     */
     private int transaction_id;
-//    @JsonProperty("amount")
+    /**
+     * Amount of money involved in the transaction.
+     */
     private double amount;
-
-//    @JsonProperty("Currency")
+    /**
+     * Currency used for the transaction.
+     */
     private String Currency;
-//    @JsonProperty("type")
+    /**
+     * Type of the transaction (e.g., credit, debit).
+     */
     private String type;
+    /**
+     * Account ID of the source of funds.
+     */
     private int from;
+    /**
+     * Account ID of the destination for funds.
+     */
     private int to;
 
+    /**
+     * Timestamp indicating when the transaction occurred.
+     */
     private Date timestamp;
 
+    /**
+     * Overridden toString method for better representation of the Transaction object.
+     */
     @Override
     public String toString() {
         return "Transaction{" +
@@ -34,13 +56,21 @@ public class Transaction {
                 '}';
     }
 
+    /**
+     * Default constructor for the Transaction class.
+     */
     public Transaction() {
     }
-
+    /**
+     * Getter for the transaction timestamp.
+     */
     public Date getTimestamp() {
         return timestamp;
     }
 
+    /**
+     * Setter for the transaction timestamp.
+     */
     public void setTimestamp(Date timestamp) {
         this.timestamp = timestamp;
     }
@@ -53,42 +83,70 @@ public class Transaction {
         this.transaction_id = transaction_id;
     }
 
+    /**
+     * Getter for the amount.
+     */
     public double getAmount() {
         return amount;
     }
-
+    /**
+     * Setter for the amount.
+     */
     public void setAmount(double amount) {
         this.amount = amount;
     }
 
+    /**
+     * Getter for the currency.
+     */
     public String getCurrency() {
         return Currency;
     }
 
+    /**
+     * Setter for the currency.
+     */
     public void setCurrency(String Currency) {
         this.Currency = Currency;
     }
 
+    /**
+     * Getter for the type of transaction
+     */
     public String getType() {
         return type;
     }
 
+    /**
+     * Setter for type of transaction
+     */
     public void setType(String type) {
         this.type = type;
     }
 
+    /**
+     * Getter for the source account id
+     */
     public int getFrom() {
         return from;
     }
 
+    /**
+     * Setter for the source account id
+     */
     public void setFrom(int from) {
         this.from = from;
     }
-
+    /**
+     * Getter for the destination account id
+     */
     public int getTo() {
         return to;
     }
 
+    /**
+     * Setter for the destination account id
+     */
     public void setTo(int to) {
         this.to = to;
     }
