+++ src/main/java/com/kirana/Kirana_Register/filter/JwtAuthFilter.java
@@ -8,7 +8,6 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-//import org.aspectj.weaver.patterns.IToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -19,6 +18,12 @@ import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
 
+/**
+ * Custom filter for handling JWT-based authentication.
+ * This filter extracts the JWT token from the request header,
+ * validates it, and sets the authenticated user in the SecurityContext.
+ * It also incorporates rate limiting using the Bucket4J library.
+ */
 @Component
 public class JwtAuthFilter extends OncePerRequestFilter {
 
@@ -31,12 +36,28 @@ public class JwtAuthFilter extends OncePerRequestFilter {
     @Autowired
     private final RateLimitConfig rateLimitConfig;
 
+    /**
+     * Constructor for JwtAuthFilter.
+     *
+     * @param jwtService         Service for JWT-related operations.
+     * @param userDetailsService Service for retrieving user details.
+     * @param rateLimitConfig    Configuration for rate limiting.
+     */
     public JwtAuthFilter(JwtService jwtService, UserInfoUserDetailsService userDetailsService, RateLimitConfig rateLimitConfig) {
         this.jwtService = jwtService;
         this.userDetailsService = userDetailsService;
         this.rateLimitConfig = rateLimitConfig;
     }
 
+    /**
+     * Filters incoming requests and processes JWT-based authentication.
+     *
+     * @param request     The HTTP request.
+     * @param response    The HTTP response.
+     * @param filterChain The filter chain.
+     * @throws ServletException If an exception occurs during the filter execution.
+     * @throws IOException      If an I/O error occurs during the filter execution.
+     */
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
         String authHeader = request.getHeader("Authorization");
