+++ src/main/java/com/kirana/Kirana_Register/repository/UserInfoRepository.java
@@ -6,6 +6,15 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import java.util.Optional;
 
 
+/**
+ * Repository interface for managing user information in MongoDB.
+ */
 public interface UserInfoRepository extends MongoRepository<UserInfo, String> {
+    /**
+     * Find a user by their username.
+     *
+     * @param username The username to search for.
+     * @return An optional containing the user information if found.
+     */
     Optional<UserInfo> findByName(String username);
 }
