+++ src/main/java/com/kirana/Kirana_Register/services/reportgeneratorservices/WeeklyReportGenerator.java
@@ -7,16 +7,33 @@ import org.springframework.stereotype.Service;
 import java.util.Date;
 import java.util.List;
 
+/**
+ * Service class responsible for generating weekly reports based on transactions.
+ */
 @Service
 public class WeeklyReportGenerator {
 
+    /**
+     * Repository for interacting with transaction information stored in the database.
+     */
     private final TransactionRepository transactionRepository;
 
+    /**
+     * Constructor for initializing the WeeklyReportGenerator service with dependencies.
+     *
+     * @param transactionRepository Repository for interacting with transaction information stored in the database.
+     */
     @Autowired
     public WeeklyReportGenerator(TransactionRepository transactionRepository) {
         this.transactionRepository = transactionRepository;
     }
 
+    /**
+     * Generates a weekly report based on transactions that occurred up to a given timestamp.
+     *
+     * @param timestamp The timestamp up to which transactions will be considered for the report.
+     * @return A WeeklyReport object containing total credit, total debit, and net balance for the specified week.
+     */
     public WeeklyReport generateWeeklyReport(Date timestamp) {
         // Calculate one week before the given timestamp
         Date startDate = new Date(timestamp.getTime() - (7 * 24 * 60 * 60 * 1000));
@@ -41,11 +58,5 @@ public class WeeklyReportGenerator {
         // Calculate net balance
         double netBalance = totalCredit - totalDebit;
         return new WeeklyReport(totalCredit, totalDebit, netBalance);
-//        Map<String, Double> weeklyReport = new HashMap<>();
-//        weeklyReport.put("totalCredit", totalCredit);
-//        weeklyReport.put("totalDebit", totalDebit);
-//        weeklyReport.put("netBalance", netBalance);
-//
-//        return weeklyReport;
     }
 }
