+++ src/main/resources/application.properties
@@ -2,6 +2,3 @@ spring.data.mongodb.uri= mongodb://localhost:27017
 spring.data.mongodb.database= kirana_register
 api.fxratesapi.url=https://api.fxratesapi.com/latest
 spring.main.allow-bean-definition-overriding=true
-
-
-#mongodb+srv://rohann:rohan@kirana.dnac4nb.mongodb.net/?retryWrites=true&w=majority
